pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]

[project]
name = "BasicTS"
version = "1.0.0"
description = "A Fair and Scalable Time Series Analysis Toolkit and benchmark."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
dependencies = [
    "easy-torch==1.3.3",
    "easydict",
    "packaging",
    "setproctitle",
    "pandas",
    "scikit-learn",
    "tables",
    "sympy",
    "openpyxl",
    "setuptools==59.5.0",
    "numpy==1.24.4",
    "tqdm==4.67.1",
    "tensorboard==2.18.0",
    "transformers==4.40.1"
]

[project.urls]
Homepage = "https://github.com/GestaltCogTeam/BasicTS"
Repository = "https://github.com/GestaltCogTeam/BasicTS"

[tool.setuptools]
package-dir = {"" = "src"}